+++ app/src/main/java/com/replica/replicaisland/AnimationComponent.kt
@@ -17,11 +17,12 @@
 
 package com.replica.replicaisland
 
-import com.replica.replicaisland.CollisionParameters.HitType
+import com.replica.replicaisland.mechanics.CollisionParameters.HitType
 import com.replica.replicaisland.core.GameObject.ActionType
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.entities.PlayerComponent
+import com.replica.replicaisland.mechanics.HotSpotSystem
 import com.replica.replicaisland.sound.SoundSystem
 import com.replica.replicaisland.sound.SoundSystem.Sound
 import kotlin.math.abs
