+++ app/src/main/java/com/replica/replicaisland/AnimationFrame.kt
@@ -15,6 +15,8 @@
  */
 package com.replica.replicaisland
 
+import com.replica.replicaisland.mechanics.CollisionVolume
+
 /**
  * A single animation frame.  Frames contain a texture, a hold time, and collision volumes to
  * use for "attacking" or "vulnerability."  This allows animated sprites to cheaply interact with
