+++ app/src/main/java/com/replica/replicaisland/GameObjectCollisionSystem.kt
@@ -17,10 +17,12 @@
 
 package com.replica.replicaisland
 
-import com.replica.replicaisland.CollisionParameters.HitType
-import com.replica.replicaisland.CollisionVolume.FlipInfo
+import com.replica.replicaisland.mechanics.CollisionParameters.HitType
+import com.replica.replicaisland.mechanics.CollisionVolume.FlipInfo
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.mechanics.AABoxCollisionVolume
+import com.replica.replicaisland.mechanics.CollisionVolume
 import java.util.*
 
 /**
@@ -169,12 +171,12 @@ class GameObjectCollisionSystem : BaseObject() {
      * or HitType.INVALID if no intersections are found.
      */
     private fun testAttackAgainstVulnerability(
-            attackVolumes: FixedSizeArray<CollisionVolume>?,
-            vulnerabilityVolumes: FixedSizeArray<CollisionVolume>?,
-            attackPosition: Vector2?,
-            vulnerabilityPosition: Vector2?,
-            attackFlip: FlipInfo,
-            vulnerabilityFlip: FlipInfo): Int {
+        attackVolumes: FixedSizeArray<CollisionVolume>?,
+        vulnerabilityVolumes: FixedSizeArray<CollisionVolume>?,
+        attackPosition: Vector2?,
+        vulnerabilityPosition: Vector2?,
+        attackFlip: FlipInfo,
+        vulnerabilityFlip: FlipInfo): Int {
         var intersectionType = HitType.INVALID
         if (attackVolumes != null && vulnerabilityVolumes != null) {
             val attackCount = attackVolumes.count
