+++ app/src/main/java/com/replica/replicaisland/HitPlayerComponent.kt
@@ -17,6 +17,7 @@ package com.replica.replicaisland
 
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.mechanics.CollisionParameters
 
 class HitPlayerComponent : GameComponent() {
     private var mDistance2 = 0f
