+++ app/src/main/java/com/replica/replicaisland/HitReactionComponent.kt
@@ -17,7 +17,7 @@
 
 package com.replica.replicaisland
 
-import com.replica.replicaisland.CollisionParameters.HitType
+import com.replica.replicaisland.mechanics.CollisionParameters.HitType
 import com.replica.replicaisland.core.GameObject.ActionType
 import com.replica.replicaisland.core.GameObject.Team
 import com.replica.replicaisland.core.GameObjectFactory.GameObjectType
