+++ app/src/main/java/com/replica/replicaisland/OrbitalMagnetComponent.kt
@@ -15,9 +15,10 @@
  */
 package com.replica.replicaisland
 
-//import com.replica.replicaisland.GravityComponent
+//import com.replica.replicaisland.mechanics.GravityComponent
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
+import com.replica.replicaisland.mechanics.GravityComponent
 import kotlin.math.sqrt
 
 class OrbitalMagnetComponent : GameComponent() {
