+++ app/src/main/java/com/replica/replicaisland/PatrolComponent.kt
@@ -18,7 +18,7 @@
 package com.replica.replicaisland
 
 import com.replica.replicaisland.core.GameObject.ActionType
-import com.replica.replicaisland.HotSpotSystem.HotSpotType
+import com.replica.replicaisland.mechanics.HotSpotSystem.HotSpotType
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import kotlin.math.abs
