+++ app/src/main/java/com/replica/replicaisland/SpriteComponent.kt
@@ -20,6 +20,7 @@ package com.replica.replicaisland
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 import com.replica.replicaisland.core.PhasedObjectManager
+import com.replica.replicaisland.mechanics.DynamicCollisionComponent
 import com.replica.replicaisland.rendering.RenderComponent
 
 /**
