+++ app/src/main/java/com/replica/replicaisland/core/GameObjectFactory.kt
@@ -1,46 +1,44 @@
 package com.replica.replicaisland.core
 
-import com.replica.replicaisland.AABoxCollisionVolume
+import com.replica.replicaisland.mechanics.AABoxCollisionVolume
 import com.replica.replicaisland.AnimationComponent
 import com.replica.replicaisland.AnimationFrame
 import com.replica.replicaisland.AnimationPlayerActivity
 import com.replica.replicaisland.AttackAtDistanceComponent
-import com.replica.replicaisland.BackgroundCollisionComponent
+import com.replica.replicaisland.mechanics.BackgroundCollisionComponent
 import com.replica.replicaisland.entities.ButtonAnimationComponent
 import com.replica.replicaisland.ChangeComponentsComponent
-import com.replica.replicaisland.ChannelSystem
-import com.replica.replicaisland.CollisionParameters
-import com.replica.replicaisland.CollisionVolume
+import com.replica.replicaisland.mechanics.ChannelSystem
+import com.replica.replicaisland.mechanics.CollisionParameters
+import com.replica.replicaisland.mechanics.CollisionVolume
 import com.replica.replicaisland.entities.CrusherAndouComponent
 import com.replica.replicaisland.DebugLog
 import com.replica.replicaisland.entities.DoorAnimationComponent
 import com.replica.replicaisland.DrawableBitmap
-import com.replica.replicaisland.DynamicCollisionComponent
+import com.replica.replicaisland.mechanics.DynamicCollisionComponent
 import com.replica.replicaisland.entities.EnemyAnimationComponent
-import com.replica.replicaisland.EventRecorder
+import com.replica.replicaisland.mechanics.EventRecorder
 import com.replica.replicaisland.FadeDrawableComponent
 import com.replica.replicaisland.FixedAnimationComponent
 import com.replica.replicaisland.FixedSizeArray
 import com.replica.replicaisland.FrameRateWatcherComponent
 import com.replica.replicaisland.GameComponent
-import com.replica.replicaisland.core.GameComponentPool
-import com.replica.replicaisland.GameFlowEvent
+import com.replica.replicaisland.mechanics.GameFlowEvent
 import com.replica.replicaisland.GenericAnimationComponent
 import com.replica.replicaisland.GhostComponent
-import com.replica.replicaisland.GravityComponent
+import com.replica.replicaisland.mechanics.GravityComponent
 import com.replica.replicaisland.HitPlayerComponent
 import com.replica.replicaisland.HitReactionComponent
 import com.replica.replicaisland.InventoryComponent
 import com.replica.replicaisland.LaunchProjectileComponent
 import com.replica.replicaisland.entities.LauncherComponent
-import com.replica.replicaisland.core.LifetimeComponent
 import com.replica.replicaisland.MotionBlurComponent
-import com.replica.replicaisland.MovementComponent
+import com.replica.replicaisland.mechanics.MovementComponent
 import com.replica.replicaisland.entities.NPCAnimationComponent
 import com.replica.replicaisland.entities.NPCComponent
 import com.replica.replicaisland.OrbitalMagnetComponent
 import com.replica.replicaisland.PatrolComponent
-import com.replica.replicaisland.PhysicsComponent
+import com.replica.replicaisland.mechanics.PhysicsComponent
 import com.replica.replicaisland.PlaySingleSoundComponent
 import com.replica.replicaisland.entities.PlayerComponent
 import com.replica.replicaisland.PopOutComponent
@@ -48,11 +46,11 @@ import com.replica.replicaisland.R
 import com.replica.replicaisland.ScrollerComponent
 import com.replica.replicaisland.SelectDialogComponent
 import com.replica.replicaisland.SimpleCollisionComponent
-import com.replica.replicaisland.SimplePhysicsComponent
+import com.replica.replicaisland.mechanics.SimplePhysicsComponent
 import com.replica.replicaisland.SleeperComponent
 import com.replica.replicaisland.SolidSurfaceComponent
 import com.replica.replicaisland.SortConstants
-import com.replica.replicaisland.SphereCollisionVolume
+import com.replica.replicaisland.mechanics.SphereCollisionVolume
 import com.replica.replicaisland.SpriteAnimation
 import com.replica.replicaisland.SpriteComponent
 import com.replica.replicaisland.TObjectPool
