+++ app/src/main/java/com/replica/replicaisland/core/LifetimeComponent.kt
@@ -2,7 +2,7 @@ package com.replica.replicaisland.core
 
 import com.replica.replicaisland.GameComponent
 import com.replica.replicaisland.GhostComponent
-import com.replica.replicaisland.HotSpotSystem
+import com.replica.replicaisland.mechanics.HotSpotSystem
 import com.replica.replicaisland.LaunchProjectileComponent
 import com.replica.replicaisland.Vector2
 import com.replica.replicaisland.sound.SoundSystem
