+++ app/src/main/java/com/replica/replicaisland/entities/DoorAnimationComponent.kt
@@ -1,6 +1,6 @@
 package com.replica.replicaisland.entities
 
-import com.replica.replicaisland.ChannelSystem
+import com.replica.replicaisland.mechanics.ChannelSystem
 import com.replica.replicaisland.GameComponent
 import com.replica.replicaisland.SolidSurfaceComponent
 import com.replica.replicaisland.SpriteComponent
