+++ app/src/main/java/com/replica/replicaisland/entities/NPCAnimationComponent.kt
@@ -1,6 +1,6 @@
 package com.replica.replicaisland.entities
 
-import com.replica.replicaisland.ChannelSystem
+import com.replica.replicaisland.mechanics.ChannelSystem
 import com.replica.replicaisland.GameComponent
 import com.replica.replicaisland.SpriteComponent
 import com.replica.replicaisland.core.BaseObject
