+++ app/src/main/java/com/replica/replicaisland/entities/NPCComponent.kt
@@ -1,10 +1,10 @@
 package com.replica.replicaisland.entities
 
-import com.replica.replicaisland.CollisionParameters
+import com.replica.replicaisland.mechanics.CollisionParameters
 import com.replica.replicaisland.GameComponent
-import com.replica.replicaisland.GameFlowEvent
+import com.replica.replicaisland.mechanics.GameFlowEvent
 import com.replica.replicaisland.HitReactionComponent
-import com.replica.replicaisland.HotSpotSystem
+import com.replica.replicaisland.mechanics.HotSpotSystem
 import com.replica.replicaisland.Utils
 import com.replica.replicaisland.Vector2
 import com.replica.replicaisland.core.BaseObject
