+++ app/src/main/java/com/replica/replicaisland/mechanics/AABoxCollisionVolume.kt
@@ -1,22 +1,6 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@file:Suppress("unused")
+package com.replica.replicaisland.mechanics
 
-package com.replica.replicaisland
-
+import com.replica.replicaisland.Vector2
 import kotlin.math.max
 
 /**
@@ -40,7 +24,7 @@ class AABoxCollisionVolume : CollisionVolume {
         bottomLeft = Vector2(offsetX, offsetY)
         widthHeight = Vector2(width, height)
     }
-    
+
     constructor(offsetX: Float, offsetY: Float, width: Float, height: Float,
                 hit: Int) : super(hit) {
         bottomLeft = Vector2(offsetX, offsetY)
