+++ app/src/main/java/com/replica/replicaisland/mechanics/CollisionVolume.kt
@@ -1,23 +1,8 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@file:Suppress("UnnecessaryVariable")
+package com.replica.replicaisland.mechanics
 
-package com.replica.replicaisland
-
-import com.replica.replicaisland.CollisionParameters.HitType
+import com.replica.replicaisland.AllocationGuard
+import com.replica.replicaisland.mechanics.CollisionParameters
+import com.replica.replicaisland.Vector2
 
 /**
  * CollisionVolume describes a volume (rectangle, sphere, etc) used for dynamic collision detection.
@@ -35,7 +20,7 @@ abstract class CollisionVolume : AllocationGuard {
     var hitType: Int
 
     constructor() : super() {
-        hitType = HitType.INVALID
+        hitType = CollisionParameters.HitType.INVALID
     }
 
     constructor(type: Int) : super() {
