+++ app/src/main/java/com/replica/replicaisland/mechanics/DynamicCollisionComponent.kt
@@ -1,20 +1,9 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.replica.replicaisland
+package com.replica.replicaisland.mechanics
 
+import com.replica.replicaisland.FixedSizeArray
+import com.replica.replicaisland.GameComponent
+import com.replica.replicaisland.HitReactionComponent
+import com.replica.replicaisland.Vector2
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 
@@ -34,7 +23,7 @@ class DynamicCollisionComponent : GameComponent() {
     override fun reset() {
         mAttackVolumes = null
         vulnerabilityVolumes = null
-        boundingVolume.center = Vector2.ZERO
+        boundingVolume.center = Vector2.Companion.ZERO
         boundingVolume.radius = 0.0f
         hitReactionComponent = null
     }
