+++ app/src/main/java/com/replica/replicaisland/mechanics/SimplePhysicsComponent.kt
@@ -1,20 +1,7 @@
-/*
- * Copyright (C) 2010 The Android Open Source Project
- * Copyright (C) 2025 Jim Andreas kotlin conversion
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.replica.replicaisland
+package com.replica.replicaisland.mechanics
 
+import com.replica.replicaisland.GameComponent
+import com.replica.replicaisland.Utils
 import com.replica.replicaisland.core.BaseObject
 import com.replica.replicaisland.core.GameObject
 
@@ -37,14 +24,14 @@ class SimplePhysicsComponent : GameComponent() {
         if (parentObject.touchingCeiling() && velocityY > 0.0f
                 || parentObject.touchingGround() && velocityY < 0.0f) {
             velocityY = -velocityY * mBounciness
-            if (Utils.close(velocityY, 0.0f)) {
+            if (Utils.Companion.close(velocityY, 0.0f)) {
                 velocityY = 0.0f
             }
         }
         if (parentObject.touchingRightWall() && velocityX > 0.0f
                 || parentObject.touchingLeftWall() && velocityX < 0.0f) {
             velocityX = -velocityX * mBounciness
-            if (Utils.close(velocityX, 0.0f)) {
+            if (Utils.Companion.close(velocityX, 0.0f)) {
                 velocityX = 0.0f
             }
         }
